+++ web/ui/react-app/src/pages/graph/PanelList.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { mount, shallow } from 'enzyme';
 import PanelList from './PanelList';
-import Checkbox from '../../common/Checkbox';
+import Checkbox from '../../components/Checkbox';
 import { Alert, Button } from 'reactstrap';
 import Panel, { PanelDefaultOptions } from './Panel';
 import { decodePanelOptionsFromQueryString } from '../../utils/urlParams';
